Removed validator file
